requirements.txt
git+https://github.com/Rapptz/discord.py.git@master
PyNaCl
requests